refactor: generic ContentHeading component
